+++ app/formSchema/analyst/cbc/tombstone.ts
@@ -10,10 +10,6 @@ const cbcTombstone: RJSFSchema = {
     'federalFundingSource',
   ],
   properties: {
-    projectNumber: {
-      type: 'string',
-      title: 'Project Number',
-    },
     originalProjectNumber: {
       type: 'string',
       title: 'Original Project Number',
@@ -26,19 +22,7 @@ const cbcTombstone: RJSFSchema = {
     intake: {
       type: 'string',
       title: 'Intake',
-      enum: [null, '1', '2', '3', '4', '5', '6'],
-    },
-    projectStatus: {
-      type: 'string',
-      title: 'Project Status',
-    },
-    projectTitle: {
-      type: 'string',
-      title: 'Project Title',
-    },
-    projectDescription: {
-      type: 'string',
-      title: 'Project Description',
+      enum: [null, 1, 2, 3, 4],
     },
     applicantContractualName: {
       type: 'string',
@@ -50,7 +34,7 @@ const cbcTombstone: RJSFSchema = {
     },
     eightThirtyMillionFunding: {
       type: 'string',
-      title: '$8.30 Million Funding',
+      title: '$830 Million Funding',
       oneOf: [
         { const: true, title: 'Yes' },
         { const: false, title: 'No' },